Replace recurring contribution tests with Tiered Checkout tests
