+++ support-e2e/tests/oneOffContributions.test.ts
@@ -13,7 +13,7 @@ interface TestDetails {
 
 const testsDetails: TestDetails[] = [
 	{ paymentType: 'Credit/Debit card', customAmount: '22.55' },
-	// { paymentType: 'PayPal' },
+	{ paymentType: 'PayPal' },
 ];
 
 afterEachTasks(test);
